Added new entities and some refactoring

+++ src/main/java/com/project/workshopmanagment/repository/ContactPointRepository.java
@@ -2,6 +2,7 @@ package com.project.workshopmanagment.repository;
 
 import com.project.workshopmanagment.entity.ContactPoint;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
-public interface ContactPointRepository extends JpaRepository<ContactPoint, Integer> {
+public interface ContactPointRepository extends JpaRepository<ContactPoint, Long> {
 }
