+++ src/main/java/com/project/workshopmanagment/repository/GraderRepository.java
@@ -2,6 +2,7 @@ package com.project.workshopmanagment.repository;
 
 import com.project.workshopmanagment.entity.Grader;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
-public interface GraderRepository extends JpaRepository <Grader, Integer> {
+public interface GraderRepository extends JpaRepository<Grader, Long> {
 }
