+++ src/main/java/com/project/workshopmanagment/repository/OrganizerRepository.java
@@ -2,6 +2,7 @@ package com.project.workshopmanagment.repository;
 
 import com.project.workshopmanagment.entity.Organizer;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
-public interface OrganizerRepository extends JpaRepository <Organizer, Integer> {
+public interface OrganizerRepository extends JpaRepository<Organizer, Long> {
 }
