+++ src/main/java/com/project/workshopmanagment/repository/RoleRepository.java
@@ -2,6 +2,7 @@ package com.project.workshopmanagment.repository;
 
 import com.project.workshopmanagment.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 }
